Bugfix: Thunar segfault when trying to open a directory with file to large
git-svn-id: file:///home/code/svn/xrdp/trunk@591 c7ce0e6c-3862-4c07-88d1-63b9e066504f

+++ sesman/chansrv/rdpdr/disk/rdpfs.h
@@ -46,8 +46,8 @@ struct fs_info
 	uint64_t last_write_time;
 	uint64_t last_change_time;
 	int file_attributes;
-	long file_size;
-	long allocation_size;    /* number of occuped cluster */
+	size_t file_size;
+	size_t allocation_size;    /* number of occuped cluster */
 	int nlink;
 	int delele_request;
 	int is_dir;
